Move a few layout items from the generator to utils module
