+++ openapi_spec_tools/layout/layout_generator.py
@@ -3,9 +3,7 @@ from typing import Any
 from typing import Optional
 
 from openapi_spec_tools.base_gen.utils import to_snake_case
-from openapi_spec_tools.layout.types import LayoutField
 from openapi_spec_tools.layout.types import LayoutNode
-from openapi_spec_tools.layout.types import PaginationField
 from openapi_spec_tools.layout.types import PaginationNames
 from openapi_spec_tools.layout.utils import DEFAULT_START
 from openapi_spec_tools.types import OasField
@@ -135,52 +133,3 @@ class LayoutGenerator:
                 path_node.children = sorted(path_node.children, key=lambda x: x.command)
 
         return main
-
-
-def pagination_text(pagination: PaginationNames, indent: str) -> str:
-    """Create text for the provided pagination parameters and indent."""
-    text = ""
-    if pagination.items_property:
-        text += f"{indent}{PaginationField.ITEM_PROP.value}: {pagination.items_property}\n"
-    if pagination.item_start:
-        text += f"{indent}{PaginationField.ITEM_START.value}: {pagination.item_start}\n"
-    if pagination.page_start:
-        text += f"{indent}{PaginationField.PAGE_START.value}: {pagination.page_start}\n"
-    if pagination.page_size:
-        text += f"{indent}{PaginationField.PAGE_SIZE.value}: {pagination.page_size}\n"
-    if pagination.next_header:
-        text += f"{indent}{PaginationField.NEXT_HEADER.value}: {pagination.next_header}\n"
-    if pagination.next_property:
-        text += f"{indent}{PaginationField.NEXT_PROP.value}: {pagination.next_property}\n"
-    return text
-
-
-def layout_node_text(node: LayoutNode) -> str:
-    """Create text for node, and all children."""
-    indent = "    "
-    text = f"{node.identifier}:\n"
-    text += f"{indent}{LayoutField.DESCRIPTION.value}: {node.description}\n"
-    text += f"{indent}{LayoutField.OPERATIONS.value}:\n"
-
-    for child in node.children:
-        text += f"{indent}- {LayoutField.NAME.value}: {child.command}\n"
-        flavor = LayoutField.OP_ID.value if not child.children else LayoutField.SUB_ID.value
-        text += f"{indent}  {flavor}: {child.identifier}\n"
-        if child.pagination:
-            text += f"{indent}  pagination:\n"
-            text += pagination_text(child.pagination, indent * 2)
-    text += "\n"
-
-    # recursively generate sections for sub-commands
-    sorted_subcommands = sorted(node.subcommands(), key=lambda x: x.identifier)
-    for child in sorted_subcommands:
-        text += layout_node_text(child)
-
-    return text
-
-
-def write_layout(filename: str, node: LayoutNode):
-    """Write the text from the node to the specified file."""
-    with open(filename, "w", encoding="utf-8", newline="\n") as fp:
-        fp.write(layout_node_text(node))
-
