+++ tests/layout/test_layout_utils.py
@@ -6,9 +6,11 @@ from openapi_spec_tools.layout.utils import check_pagination_definitions
 from openapi_spec_tools.layout.utils import data_to_node
 from openapi_spec_tools.layout.utils import field_to_list
 from openapi_spec_tools.layout.utils import file_to_tree
+from openapi_spec_tools.layout.utils import layout_node_text
 from openapi_spec_tools.layout.utils import open_layout
 from openapi_spec_tools.layout.utils import operation_duplicates
 from openapi_spec_tools.layout.utils import operation_order
+from openapi_spec_tools.layout.utils import pagination_text
 from openapi_spec_tools.layout.utils import parse_extras
 from openapi_spec_tools.layout.utils import parse_pagination
 from openapi_spec_tools.layout.utils import parse_to_tree
@@ -525,3 +527,82 @@ def test_file_to_tree() -> None:
     tree = file_to_tree(filename, "owners")
     assert "owners" == tree.command
     assert set() == {p.command for p in tree.subcommands()}
+
+
+ALL_PAGE_TEXT = """\
+itemProperty: itemProp
+itemStart: item
+pageStart: page
+pageSize: size
+nextHeader: My-header
+nextProperty: nextProp
+"""
+ITEMS_PAGE_TEXT = """\
+  itemProperty: itemProp
+  itemStart: item
+"""
+PAGE_PAGE_TEXT = """\
+****pageStart: page
+****pageSize: size
+"""
+
+@pytest.mark.parametrize(
+    ["pagination", "indent", "expected"],
+    [
+        pytest.param(
+            PaginationNames(
+                page_size="size",
+                page_start="page",
+                item_start="item",
+                items_property="itemProp",
+                next_property="nextProp",
+                next_header="My-header",
+            ),
+            "",
+            ALL_PAGE_TEXT,
+            id="all",
+        ),
+        pytest.param(
+            PaginationNames(item_start="item", items_property="itemProp"),
+            "  ",
+            ITEMS_PAGE_TEXT,
+            id="items",
+        ),
+        pytest.param(
+            PaginationNames(page_size="size", page_start="page"),
+            "****",
+            PAGE_PAGE_TEXT,
+            id="page",
+        ),
+    ]
+)
+def test_pagination_text(pagination, indent, expected):
+    actual = pagination_text(pagination, indent)
+    assert expected == actual
+
+
+def test_layout_node_text():
+    node = LayoutNode(
+        command="parent",
+        identifier="this_is_it",
+        description="summary",
+        children=[
+            LayoutNode(command="child1", identifier="my_child", pagination=PaginationNames(page_size="page")),
+            LayoutNode(command="child2", identifier="another_op")
+        ]
+    )
+    text = layout_node_text(node)
+    assert text == """\
+this_is_it:
+    description: summary
+    operations:
+    - name: child1
+      operationId: my_child
+      pagination:
+        pageSize: page
+    - name: child2
+      operationId: another_op
+
+"""
+
+
